Wrap `TsVector::concat` in parens
The precedence of this operator treats `x @@ y || z` as `(x @@ y) || z`.
We want it the other way around. Arguably, it should be there
responsibility of `matches` to put parens around its argument, but I
don't think it matters as much. Really, we might want to consider always
emitting parens for all our operators.

Fixes #7